fix: high cardinality policy_violations metrics
The exporter initialises all possible labels of dependency_track_project_policy_violations
metrics which adds 72 metrics for each project.

https://github.com/jetstack/dependency-track-exporter/issues/53 has details of the rational but there are usecases where
it is desirable to only have non-zero values of the metrics available.

This change adds a new argument that enables only non-zero value metrics to be returned.
This flag defaults to the current behaviour so is an opt in feature.